+++ motionplan/constraint.go
@@ -43,7 +43,7 @@ func (c *constraintHandler) CheckConstraintPath(ci *ConstraintInput, resolution 
 	if err != nil {
 		return false, nil
 	}
-	steps := GetSteps(ci.StartPos, ci.EndPos, resolution)
+	steps := PathStepCount(ci.StartPos, ci.EndPos, resolution)
 
 	var lastGood []referenceframe.Input
 	// Seed with just the start position to walk the path
